LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-getComputedStyle-001-expected.txt


FAIL Basic case assert_equals: expected "40px" but got "0px"
FAIL Mixed writing modes and directions assert_equals: expected "40px" but got "0px"
FAIL With containing block padding assert_equals: expected "40px" but got "0px"
